contoh3/requirements.txt
google-cloud-storage==1.28.1
Flask==1.1.2
mongoengine==0.20.0
scikit-learn==0.23.1
